fix: update pallet storage provider config
